Removed key overrides on FETCH_ALL_MEMBERS
You dont want to remove all data on the start of an action, given possible longer loading times, it would be an eye sore to see a message stating there are no users then immediately disappear. It should be seamless. As well as removing the team data on a failure. Imagine you try to refresh and there was some error, but now it says you have no members. Regardless of the error, it would be an issue from a user experience standpoint.